refactor multigrid DFT code
