+++ pyscf/pbc/gto/cell.py
@@ -1361,6 +1361,10 @@ class Cell(mole.Mole):
             infinity vacuum (inf_vacuum) or truncated Coulomb potential
             (analytic_2d_1). Unless explicitly specified, analytic_2d_1 is
             used for 2D system and inf_vacuum is assumed for 1D and 0D.
+        rcut_by_shell_radius : bool
+            If True, radius cutoff is determined by shell radius;
+            otherwise, it is determined by overlap integral.
+            Default value is False;
 
     (See other attributes in :class:`Mole`)
 
@@ -1389,6 +1393,7 @@ class Cell(mole.Mole):
         #       density-fitting class.  This determines how the ewald produces
         #       its energy.
         self.low_dim_ft_type = None
+        self.rcut_by_shell_radius = False
 
 ##################################################
 # These attributes are initialized by build function if not given
@@ -1681,9 +1686,11 @@ class Cell(mole.Mole):
             return self
 
         if self.rcut is None or self._rcut_from_build:
-            #self._rcut = max([self.bas_rcut(ib, self.precision)
-            #                  for ib in range(self.nbas)] + [0])
-            self._rcut = max(0, rcut_by_shells(self).max())
+            if not self.rcut_by_shell_radius:
+                self._rcut = max([self.bas_rcut(ib, self.precision)
+                                  for ib in range(self.nbas)] + [0])
+            else:
+                self._rcut = max(0, rcut_by_shells(self).max())
             self._rcut_from_build = True
 
         _a = self.lattice_vectors()
